+++ packages/content-blocks-kit/tests/Block/ButtonGroupViewTest.php
@@ -8,6 +8,7 @@ use ContentBlocks\Kit\Block\ButtonGroupBlock;
 use ContentBlocks\Kit\RichText\RichTextSanitizerFactory;
 use ContentBlocks\Kit\Twig\ChoiceTokenExtension;
 use ContentBlocks\Kit\Twig\SafeContentExtension;
+use ContentBlocks\Twig\ColorToneExtension;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\Extension\Validator\ValidatorExtension;
@@ -88,6 +89,7 @@ final class ButtonGroupViewTest extends TestCase
         $env = new Environment($loader, ['strict_variables' => true]);
         $env->addExtension(new ChoiceTokenExtension());
         $env->addExtension(new SafeContentExtension(RichTextSanitizerFactory::create()));
+        $env->addExtension(new ColorToneExtension());
 
         return $env->render('@ContentBlocksKit/block/button_group/view.html.twig', [
             'data' => $data + (new ButtonGroupBlock())->getDefaultData(),