+++ packages/content-blocks-kit/tests/Block/ChoiceFieldCoverageTest.php
@@ -11,6 +11,7 @@ use ContentBlocks\Kit\RichText\RichTextSanitizerFactory;
 use ContentBlocks\Kit\Twig\ChoiceTokenExtension;
 use ContentBlocks\Kit\Twig\IconExtension;
 use ContentBlocks\Kit\Twig\SafeContentExtension;
+use ContentBlocks\Twig\ColorToneExtension;
 use ContentBlocks\Twig\ImageExtension;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
@@ -201,6 +202,7 @@ final class ChoiceFieldCoverageTest extends TestCase
         $env = new Environment($loader, ['strict_variables' => false]);
         $env->addExtension(new ChoiceTokenExtension());
         $env->addExtension(new SafeContentExtension(RichTextSanitizerFactory::create()));
+        $env->addExtension(new ColorToneExtension());
         $env->addExtension(new IconExtension(new IconRegistry()));
         $env->addExtension(new ImageExtension(new PassthroughImageUrlResolver()));
         $env->addExtension(new TranslationExtension(new class () implements TranslatorInterface {